+++ kurtosis-lambda/impl/near_lambda.ts
@@ -3,6 +3,64 @@ import { KurtosisLambda } from "kurtosis-lambda-api-lib";
 import { Result, ok, err } from "neverthrow";
 import * as log from "loglevel";
 
+
+const TCP_PROTOCOL: string = "tcp";
+
+const WAMP_BACKEND_SECRET: string = "back";
+
+const DOCKER_PORT_PROTOCOL_SEPARATOR = "/";
+
+// Explorer WAMP
+const EXPLORER_WAMP_SERVICE_ID: ServiceID = "wamp";
+const EXPLORER_WAMP_IMAGE: string = "near-explorer_wamp";
+const EXPLORER_WAMP_PORT_NUM: number = 8080;
+const EXPLORER_WAMP_DOCKER_PORT_DESC: string = EXPLORER_WAMP_PORT_NUM.toString() + DOCKER_PORT_PROTOCOL_SEPARATOR + TCP_PROTOCOL;
+// const EXPLORER_WAMP_PORT_ENVVAR: string = "WAMP_EXPLORER_PORT";
+// const EXPLORER_WAMP_BACKEND_SECRET_ENVVAR: string = "WAMP_EXPLORER_BACKEND_SECRET";
+
+const EXPLORER_WAMP_STATIC_ENVVARS: Map<string, string> = new Map(Object.entries({
+    "WAMP_NEAR_EXPLORER_PORT": EXPLORER_WAMP_PORT_NUM.toString(),
+    "WAMP_NEAR_EXPLORER_BACKEND_SECRET": WAMP_BACKEND_SECRET,
+}));
+
+// Explorer Backend
+const EXPLORER_BACKEND_SERVICE_ID: ServiceID = "backend";
+const EXPLORER_BACKEND_IMAGE: string = "near-explorer_backend";
+const EXPLORER_BACKEND_WAMP_URL_ENVVAR: string = "WAMP_NEAR_EXPLORER_URL"
+// const EXPLORER_BACKEND_WAMP_SECRET_ENVVAR: string  = "WAMP_EXPLORER_BACKEND_SECRET";
+const EXPLORER_BACKEND_STATIC_ENVVARS: Map<string, string> = new Map(Object.entries({
+    "WAMP_NEAR_EXPLORER_BACKEND_SECRET": WAMP_BACKEND_SECRET,
+}));
+const EXPLORER_BACKEND_ENTRYPOINT_ARGS: string[] = [
+    "npm", 
+    "run", 
+    "start:testnet-with-indexer"
+];
+// TODO Mem limit??
+
+// Explorer Frontend
+const EXPLORER_FRONTEND_SERVICE_ID: ServiceID = "frontend";
+const EXPLORER_FRONTEND_IMAGE: string = "near-explorer_frontend";
+const EXPLORER_FRONTEND_PORT_NUM: number = 3000;
+const EXPLORER_FRONTEND_DOCKER_PORT_DESC: string = EXPLORER_FRONTEND_PORT_NUM.toString() + DOCKER_PORT_PROTOCOL_SEPARATOR + TCP_PROTOCOL;
+// const EXPLORER_FRONTEND_PORT_ENVVAR: string = "PORT";
+const EXPLORER_FRONTEND_WAMP_INTERNAL_URL_ENVVAR: string = "WAMP_NEAR_EXPLORER_INTERNAL_URL";
+const EXPLORER_FRONTEND_WAMP_EXTERNAL_URL_ENVVAR: string = "WAMP_NEAR_EXPLORER_URL";
+// const EXPLORER_FRONTEND_NEAR_NETWORKS_ENVVAR: string = "NEAR_NETWORKS";
+// const EXPLORER_FRONTEND_NEAR_NETWORKS_VALUE: string = "[{\"name\": \"testnet\", \"explorerLink\": \"http://localhost:3000/\", \"aliases\": [\"localhost:3000\", \"127.0.0.1:3000\"], \"nearWalletProfilePrefix\": \"https://wallet.testnet.near.org/profile\"}]";
+// const EXPLORER_FRONTEND_DATA_SOURCE_ENVVAR: string = "EXPLORER_DATA_SOURCE";
+// const EXPLORER_FRONTEND_INDEXER_DATA_SOURCE: string = "INDEXER_BACKEND";
+
+const EXPLORER_FRONTEND_STATIC_ENVVARS: Map<string, string> = new Map(Object.entries({
+    "PORT": EXPLORER_FRONTEND_PORT_NUM.toString(),
+    "EXPLORER_DATA_SOURCE": "INDEXER_BACKEND",
+    // It's not clear what this value does - it's pulled as-is from https://github.com/near/near-explorer/blob/master/frontend/package.json#L31
+    "NEAR_NETWORKS": "[{\"name\": \"testnet\", \"explorerLink\": \"http://localhost:3000/\", \"aliases\": [\"localhost:3000\", \"127.0.0.1:3000\"], \"nearWalletProfilePrefix\": \"https://wallet.testnet.near.org/profile\"}]",
+}));
+
+type ContainerRunConfigSupplier = (ipAddr: string, generatedFileFilepaths: Map<string, string>, staticFileFilepaths: Map<StaticFileID, string>) => Result<ContainerRunConfig, Error>;
+
+/*
 //Contract Helper DB
 const CONTRACT_HELPER_DB_SERVICE_ID: ServiceID = "contract-helper-db"
 const CONTRACT_HELPER_DB_IMAGE: string = "nearprotocol/contract-helper-db"
@@ -62,20 +120,18 @@ const BRIDGE_FRONTEND_SERVICE_ID: ServiceID = "bridge-frontend"
 const BRIDGE_FRONTEND_IMAGE: string = "nearprotocol/bridge-frontend"
 const BRIDGE_FRONTEND_PORT_NUMBER: string = "1234"
 const BRIDGE_FRONTEND_PROTOCOL: string = "tcp"
-
+*/
 
 interface NearLambdaParams {
 }
 
 class NearLambdaResult {
-    readonly walletURL: string
-    readonly bridgeURL: string
-    readonly explorerURL: string
+    // When Kurtosis is in debug mode, the explorer frontend's port will be bound to a port on the user's machine so they can access the frontend
+    //  even though the frontend is running inside Docker. When Kurtosis is not in debug mode, this will be null.
+    private readonly explorerHostMachineUrl: string | null
 
-    constructor(walletURL: string, bridgeURL: string, explorerURL: string) {
-        this.walletURL = walletURL;
-        this.bridgeURL = bridgeURL;
-        this.explorerURL = explorerURL;
+    constructor(explorerHostMachineUrl: string | null) {
+        this.explorerHostMachineUrl = explorerHostMachineUrl;
     }
 }
 
@@ -84,10 +140,9 @@ export class NearLambda implements KurtosisLambda {
 
     async execute(networkCtx: NetworkContext, serializedParams: string): Promise<Result<string, Error>> {
         log.info("Near Lambda receives serializedParams '" + serializedParams + "'");
-        let params: NearLambdaParams;
-
+        let args: ExecuteArgs;
         try {
-            params = JSON.parse(serializedParams)
+            args = JSON.parse(serializedParams)
         } catch (e: any) {
             // Sadly, we have to do this because there's no great way to enforce the caught thing being an error
             // See: https://stackoverflow.com/questions/30469261/checking-for-typeof-error-in-js
@@ -98,6 +153,42 @@ export class NearLambda implements KurtosisLambda {
                 "it's not an Error so we can't report any more information than this"));
         }
 
+        // TODO handle custom params here
+
+        const addWampServiceResult: Result<[ServiceContext, Map<string, PortBinding>], Error> = await NearLambda.addWampService(networkCtx)
+        if (addWampServiceResult.isErr()) {
+            return err(addWampServiceResult.error);
+        }
+        const [wampServiceCtx, wampHostPortBindings] = addWampServiceResult.value;
+        const wampHostMachinePortBindingOpt: PortBinding | undefined = wampHostPortBindings.get(EXPLORER_WAMP_DOCKER_PORT_DESC);
+
+        const addBackendServiceResult: Result<[ServiceContext, Map<string, PortBinding>], Error> = await NearLambda.addBackendService(networkCtx)
+        if (addBackendServiceResult.isErr()) {
+            return err(addBackendServiceResult.error);
+        }
+        const [backendServiceCtx, backendHostPortBindings] = addBackendServiceResult.value;
+
+        const addFrontendServiceResult: Result<[ServiceContext, Map<string, PortBinding>], Error> = await NearLambda.addFrontendService(networkCtx, wampHostMachinePortBindingOpt)
+        if (addFrontendServiceResult.isErr()) {
+            return err(addFrontendServiceResult.error);
+        }
+        const [frontendServiceCtx, frontendHostPortBindings] = addFrontendServiceResult.value;
+        const frontendHostMachinePortBindingOpt: PortBinding | undefined = frontendHostPortBindings.get(EXPLORER_FRONTEND_DOCKER_PORT_DESC);
+        let frontendHostMachineUrl: string | null;
+        if (frontendHostMachinePortBindingOpt !== undefined) {
+            const frontendHostMachinePortNumResult: Result<number, Error> = NearLambda.getPortNumFromHostMachinePortBinding(frontendHostMachinePortBindingOpt);
+            if (frontendHostMachinePortNumResult.isErr()) {
+                return err(frontendHostMachinePortNumResult.error);
+            }
+            const frontendHostMachinePortNum: number = frontendHostMachinePortNumResult.value;
+            frontendHostMachineUrl = "http://" + frontendHostMachinePortBindingOpt.getInterfaceIp() + ":" + frontendHostMachinePortNum.toString();
+        } else {
+            // There was no Docker-external host machine port bound for the frontend, so there's no way we can give a URL to the user that's accessible
+            frontendHostMachineUrl = null;
+        }
+
+
+        /*
         //Contract Helper DB
         const contractHelperDBResult: Result<[ServiceContext, Map<string, PortBinding>], Error> = await this.addContractHelperDbService(networkCtx);
         if (!contractHelperDBResult.isOk()) {
@@ -127,13 +218,11 @@ export class NearLambda implements KurtosisLambda {
         log.info(INDEXER_SERVICE_ID + "with IP: " + indexerResult.value[0].getIPAddress() + " and port bindings: " + indexerResult.value[1]);
 
         //Linkdrop
-        /*
         const linkdropResult: Result<[ServiceContext, Map<string, PortBinding>], Error> = await this.addLikdropService(networkCtx);
         if (!linkdropResult.isOk()) {
             return err(linkdropResult.error);
         }
         log.info(LINKDROP_SERVICE_ID + "with IP: " + linkdropResult.value[0].getIPAddress() + " and port bindings: " + linkdropResult.value[1]);
-        */
         
         //Explorer Wamp
         const explorerWampResult: Result<[ServiceContext, Map<string, PortBinding>], Error> = await this.addExplorerWampService(networkCtx);
@@ -169,12 +258,9 @@ export class NearLambda implements KurtosisLambda {
             return err(bridgeFrontendResult.error);
         }
         log.info(BRIDGE_SERVICE_ID + "with IP: " + bridgeFrontendResult.value[0].getIPAddress() + " and port bindings: " + bridgeFrontendResult.value[1]);
+        */
 
-        const nearLambdaResult: NearLambdaResult = new NearLambdaResult(
-            "http://localhost:"+ walletResult.value[1].get(WALLET_PORT_NUMBER + "/" + WALLET_PROTOCOL)!.getInterfacePort(),
-            "http://localhost:"+ bridgeResult.value[1].get(BRIDGE_PORT_NUMBER_01 + "/" + BRIDGE_PROTOCOL)!.getInterfacePort(),
-            "http://localhost:"+ explorerFrontendResult.value[1].get(EXPLORER_FRONTEND_PORT_NUMBER + "/" + EXPLORER_FRONTEND_PROTOCOL)!.getInterfacePort()
-        );
+        const nearLambdaResult: NearLambdaResult = new NearLambdaResult(frontendHostMachineUrl);
 
         let stringResult;
         try {
@@ -196,7 +282,108 @@ export class NearLambda implements KurtosisLambda {
     // ====================================================================================================
     //                                       Private helper functions
     // ====================================================================================================
+    private static async addWampService(networkCtx: NetworkContext): Promise<Result<[ServiceContext, Map<string, PortBinding>], Error>> {
+        const usedPortsSet: Set<string> = new Set();
+        usedPortsSet.add(EXPLORER_WAMP_DOCKER_PORT_DESC)
+        const containerCreationConfig: ContainerCreationConfig = new ContainerCreationConfigBuilder(
+            EXPLORER_WAMP_IMAGE,
+        ).withUsedPorts(
+            usedPortsSet,
+        ).build();
 
+        const envVars: Map<string, string> = new Map(EXPLORER_WAMP_STATIC_ENVVARS);
+        const containerRunConfigSupplier: ContainerRunConfigSupplier = (ipAddr: string, generatedFileFilepaths: Map<string, string>, staticFileFilepaths: Map<StaticFileID, string>) => {
+            const result: ContainerRunConfig = new ContainerRunConfigBuilder().withEnvironmentVariableOverrides(
+                envVars
+            ).build();
+            return ok(result);
+        }
+        
+        const addServiceResult: Result<[ServiceContext, Map<string, PortBinding>], Error> = await networkCtx.addService(EXPLORER_WAMP_SERVICE_ID, containerCreationConfig, containerRunConfigSupplier);
+        if (addServiceResult.isErr()) {
+            return err(addServiceResult.error);
+        }
+        return ok(addServiceResult.value);
+    }
+
+    private static async addBackendService(networkCtx: NetworkContext): Promise<Result<[ServiceContext, Map<string, PortBinding>], Error>> {
+        const containerCreationConfig: ContainerCreationConfig = new ContainerCreationConfigBuilder(
+            EXPLORER_BACKEND_IMAGE,
+        ).build();
+
+        const envVars: Map<string, string> = new Map(EXPLORER_BACKEND_STATIC_ENVVARS);
+        envVars.set(EXPLORER_BACKEND_WAMP_URL_ENVVAR, "ws://" + EXPLORER_WAMP_SERVICE_ID + ":" + EXPLORER_WAMP_PORT_NUM + "/ws");
+        // envVars.set(EXPLORER_BACKEND_WAMP_SECRET_ENVVAR, WAMP_BACKEND_SECRET);
+        const containerRunConfigSupplier: ContainerRunConfigSupplier = (ipAddr: string, generatedFileFilepaths: Map<string, string>, staticFileFilepaths: Map<StaticFileID, string>) => {
+            const result: ContainerRunConfig = new ContainerRunConfigBuilder().withEnvironmentVariableOverrides(
+                envVars
+            ).wthEntrypointOverride(
+                EXPLORER_BACKEND_ENTRYPOINT_ARGS
+            ).build();
+            return ok(result);
+        }
+        
+        const addServiceResult: Result<[ServiceContext, Map<string, PortBinding>], Error> = await networkCtx.addService(EXPLORER_BACKEND_SERVICE_ID, containerCreationConfig, containerRunConfigSupplier);
+        if (addServiceResult.isErr()) {
+            return err(addServiceResult.error);
+        }
+        return ok(addServiceResult.value);
+    }
+
+    private static async addFrontendService(networkCtx: NetworkContext, wampHostMachinePortBindingOpt?: PortBinding): Promise<Result<[ServiceContext, Map<string, PortBinding>], Error>> {
+        const usedPortsSet: Set<string> = new Set();
+        usedPortsSet.add(EXPLORER_FRONTEND_DOCKER_PORT_DESC)
+        const containerCreationConfig: ContainerCreationConfig = new ContainerCreationConfigBuilder(
+            EXPLORER_FRONTEND_IMAGE,
+        ).withUsedPorts(
+            usedPortsSet,
+        ).build();
+
+
+        const envVars: Map<string, string> = new Map(EXPLORER_FRONTEND_STATIC_ENVVARS);
+        // envVars.set(EXPLORER_FRONTEND_PORT_ENVVAR, EXPLORER_FRONTEND_PORT_NUM.toString());
+        // envVars.set(EXPLORER_FRONTEND_NEAR_NETWORKS_ENVVAR, EXPLORER_FRONTEND_NEAR_NETWORKS_VALUE);
+        // envVars.set(EXPLORER_FRONTEND_DATA_SOURCE_ENVVAR, "INDEXER_BACKEND");
+        envVars.set(EXPLORER_FRONTEND_WAMP_INTERNAL_URL_ENVVAR, "ws://" + EXPLORER_WAMP_SERVICE_ID + ":" + EXPLORER_WAMP_PORT_NUM + "/ws");
+        // If there's no host machine WAMP port (i.e. Kurtosis isn't running in debug mode) then we can't set the WAMP Docker-external variable
+        if (wampHostMachinePortBindingOpt !== undefined) {
+            const wampHostMachinePortNumResult: Result<number, Error> = NearLambda.getPortNumFromHostMachinePortBinding(wampHostMachinePortBindingOpt);
+            if (wampHostMachinePortNumResult.isErr()) {
+                return err(wampHostMachinePortNumResult.error);
+            }
+            const wampHostMachinePortNum: number = wampHostMachinePortNumResult.value;
+            const wampHostMachineInterfaceIp: string = wampHostMachinePortBindingOpt.getInterfaceIp();
+            envVars.set(EXPLORER_FRONTEND_WAMP_EXTERNAL_URL_ENVVAR, "ws://" + wampHostMachineInterfaceIp + ":" + wampHostMachinePortNum.toString() + "/ws");  // This is the WS port on the user's local machine, which we'll only know when we start the WAMP service
+        }
+        const containerRunConfigSupplier: ContainerRunConfigSupplier = (ipAddr: string, generatedFileFilepaths: Map<string, string>, staticFileFilepaths: Map<StaticFileID, string>) => {
+            const result: ContainerRunConfig = new ContainerRunConfigBuilder().withEnvironmentVariableOverrides(
+                envVars
+            ).build();
+            return ok(result);
+        }
+        
+        const addServiceResult: Result<[ServiceContext, Map<string, PortBinding>], Error> = await networkCtx.addService(EXPLORER_FRONTEND_SERVICE_ID, containerCreationConfig, containerRunConfigSupplier);
+        if (addServiceResult.isErr()) {
+            return err(addServiceResult.error);
+        }
+        return ok(addServiceResult.value);
+    }
+
+    private static getPortNumFromHostMachinePortBinding(binding: PortBinding): Result<number, Error> {
+        const portStr: string = binding.getInterfacePort();
+        const portStrParts: string[] = portStr.split(DOCKER_PORT_PROTOCOL_SEPARATOR);
+        const portNumStr: string = portStrParts[0];
+        let portNum: number;
+        try {
+            portNum = parseInt(portNumStr);
+        } catch (e: any) {
+            return err(new Error("Couldn't parse host machine port binding number string '" + portNumStr + "' to a number"));
+        }
+        return ok(portNum);
+    }
+
+
+    /*
     private async addContractHelperDbService(networkCtx: NetworkContext): Promise<Result<[ServiceContext, Map<string, PortBinding>], Error>> {
         const usedPortsSet: Set<string> = new Set();
         const containerCreationConfig: ContainerCreationConfig = new ContainerCreationConfigBuilder(
@@ -386,8 +573,8 @@ export class NearLambda implements KurtosisLambda {
         (ipAddr: string, generatedFileFilepaths: Map<string, string>, staticFileFilepaths: Map<StaticFileID, string>) => {
             const environmentVariables: Map<string, string> = new Map();
             const result: ContainerRunConfig = new ContainerRunConfigBuilder().withEnvironmentVariableOverrides(
-                environmentVariables.set("WAMP_NEAR_EXPLORER_PORT", EXPLORER_WAMP_PORT_NUMBER_01)
-                .set("WAMP_NEAR_EXPLORER_BACKEND_SECRET", "back")
+                environmentVariables.set("WAMP_EXPLORER_PORT", EXPLORER_WAMP_PORT_NUMBER_01)
+                .set("WAMP_EXPLORER_BACKEND_SECRET", "back")
             ).build()
             return ok(result);
         }
@@ -410,8 +597,8 @@ export class NearLambda implements KurtosisLambda {
             const result: ContainerRunConfig = new ContainerRunConfigBuilder().withEnvironmentVariableOverrides(
                 environmentVariables.set("NEAR_RPC_URL", "http://" + BRIDGE_SERVICE_ID + ":" + BRIDGE_PORT_NUMBER_01)
                 .set("NEAR_GENESIS_RECORDS_URL", "http://" + BRIDGE_SERVICE_ID + ":" + BRIDGE_PORT_NUMBER_05 + "/genesis.json")
-                .set("WAMP_NEAR_EXPLORER_URL", "ws://" + EXPLORER_WAMP_SERVICE_ID + ":" + EXPLORER_WAMP_PORT_NUMBER_01  + "/ws")
-                .set("WAMP_NEAR_EXPLORER_BACKEND_SECRET", "back")
+                .set("WAMP_EXPLORER_URL", "ws://" + EXPLORER_WAMP_SERVICE_ID + ":" + EXPLORER_WAMP_PORT_NUMBER_01  + "/ws")
+                .set("WAMP_EXPLORER_BACKEND_SECRET", "back")
             ).wthEntrypointOverride([
                 "/bin/sh",
                 "-c",
@@ -461,8 +648,8 @@ export class NearLambda implements KurtosisLambda {
             const environmentVariables: Map<string, string> = new Map();
             const result: ContainerRunConfig = new ContainerRunConfigBuilder().withEnvironmentVariableOverrides(
                 environmentVariables.set("PORT", EXPLORER_FRONTEND_PORT_NUMBER)
-                .set("WAMP_NEAR_EXPLORER_URL", "ws://localhost:" + explorerWampHostPort + "/ws")
-                .set("WAMP_NEAR_EXPLORER_INTERNAL_URL", "ws://" + EXPLORER_WAMP_SERVICE_ID + ":" + EXPLORER_WAMP_PORT_NUMBER_01 + "/ws")
+                .set("WAMP_EXPLORER_URL", "ws://localhost:" + explorerWampHostPort + "/ws")
+                .set("WAMP_EXPLORER_INTERNAL_URL", "ws://" + EXPLORER_WAMP_SERVICE_ID + ":" + EXPLORER_WAMP_PORT_NUMBER_01 + "/ws")
                 .set("NEAR_NETWORKS", "[{\"name\": \"localhostnet\",\"explorerLink\": \"http://localhost:" + contractHelperAppHostPort + "\",\"aliases\": [\"localhost:" + contractHelperAppHostPort + "\", \"localhost\", \"127.0.0.1\", \"127.0.0.1:" + contractHelperAppHostPort  + "\"],\"nearWalletProfilePrefix\": \"http://localhost:" + walletHostPort + "/profile\"}]")
             ).build()
             return ok(result);
@@ -519,5 +706,6 @@ export class NearLambda implements KurtosisLambda {
         return addServiceResult
 
     }
+    */
 
 }