+++ docs/changelog.md
@@ -1,4 +1,6 @@
 # TBD
+### Features
+* Start a NEAR Explorer backed by a local indexer backed by the bridge, use an Explorer backed by the publicly-available testnet indexer
 
 # 0.1.0
 * Initial commit